add placeholder test for NumericTextField #2244
