+++ src/BlazorUI/Bit.BlazorUI.Tests/NumericTextField/BitNumericTextFieldGeneralTests.cs
@@ -1,0 +1,29 @@
+﻿using Bunit;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Bit.BlazorUI.Tests.NumericTextField;
+
+[TestClass]
+public class BitNumericTextFieldGeneralTests : BunitTestContext
+{
+    [TestInitialize]
+    public void SetupJsInteropMode()
+    {
+        Context.JSInterop.Mode = JSRuntimeMode.Loose;
+    }
+
+    [DataTestMethod,
+     DataRow("The placeholder")]
+    public void BitNumericTextFieldShouldHaveCorrectPlaceholder(string placeholder)
+    {
+        var component = RenderComponent<BitNumericTextField<byte>>(parameters =>
+        {
+            parameters.Add(p => p.Placeholder, placeholder);
+        });
+
+        var numericTextFieldPlaceholder = component.Find(".bit-ntf-input");
+
+        Assert.IsTrue(numericTextFieldPlaceholder.HasAttribute("placeholder"));
+        Assert.AreEqual(numericTextFieldPlaceholder.GetAttribute("placeholder"), placeholder);
+    }
+}
