+++ src/BlazorUI/Bit.BlazorUI.Tests/NumericTextField/Double/BitNumericTextFieldDoubleTests.cs
@@ -21,21 +21,6 @@ public class BitNumericTextFieldDoubleTests : BunitTestContext
     }
 
     [DataTestMethod,
-     DataRow("The placeholder")]
-    public void BitNumericTextFieldShouldHaveCorrectPlaceholder(string placeholder)
-    {
-        var component = RenderComponent<BitNumericTextField<byte>>(parameters =>
-        {
-            parameters.Add(p => p.Placeholder, placeholder);
-        });
-
-        var numericTextFieldPlaceholder = component.Find(".bit-ntf-input");
-
-        Assert.IsTrue(numericTextFieldPlaceholder.HasAttribute("placeholder"));
-        Assert.AreEqual(numericTextFieldPlaceholder.GetAttribute("placeholder"), placeholder);
-    }
-
-    [DataTestMethod,
      DataRow(null),
      DataRow("The Label")
     ]
